requirement.txt
ipykernel

# Environment Variable Management
python-dotenv

# Langchain Core
langchain
langchain-core

# Open-AI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# huggingface Integration
langchain-huggingface
transformers
huggingface-hub
torch

# machine learning utilities
numpy
scikit-learn

# UI
streamlit